Use ints for ResponseLog data

+++ DevilDaggersWebsite/Entities/ResponseLog.cs
@@ -6,12 +6,12 @@ namespace DevilDaggersWebsite.Entities
 	public class ResponseLog
 	{
 		[Key]
-		public long Id { get; set; }
+		public int Id { get; set; }
 
 		[StringLength(63)]
 		public string Path { get; set; } = null!;
 
-		public long ResponseTimeMicroseconds { get; set; }
+		public int ResponseTimeMicroseconds { get; set; }
 
 		public DateTime DateTime { get; set; }
 	}
